pyproject.toml
[project]
name = "o3-deep-research-with-bing-search"
version = "0.1.0"
description = "O3 Deep Research with Bing Search using Azure OpenAI"
authors = [
    {name = "Yu-Hong Lin", email = "your.email@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
